Перенесла функции генерации комментария и фотографии в другой файл
